backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "dj-database-url>=3.0.1",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-crontask>=1.0.3",
    "django-ninja>=1.5.1",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "gunicorn>=23.0.0",
    "openai>=2.11.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "resend>=2.19.0",
]
